chore: accept new Cruft update

+++ move_template_files.sh
@@ -1,4 +1,9 @@
 #!/bin/sh
 
-rsync -avh nomad-parser-yambospectra/ .
-rm -rfv nomad-parser-yambospectra
+if ! command -v rsync &> /dev/null; then
+  echo "rsync required, but not installed!"
+  exit 1
+else
+  rsync -avh nomad-parser-yambospectra/ .
+  rm -rfv nomad-parser-yambospectra
+fi
